Finished RollOptions and test
