+++ package.json
@@ -18,6 +18,7 @@
   "license": "MIT",
   "devDependencies": {
     "chai": "^3.5.0",
+    "eslint": "^4.7.0",
     "mocha": "^2.4.5"
   }
 }
