+++ test/RollOptions.spec.js
@@ -1,7 +1,7 @@
 /* global describe, it */
 const chai = require('chai');
 const expect = chai.expect;
-const RollOptions = require('../src/RO');
+const RollOptions = require('../src/RollOptions');
 
 describe('RollOptions', function() {
   it('should accept an empty string', function(){
@@ -98,6 +98,138 @@ describe('RollOptions', function() {
       expect(options.needReroll(3)).to.be.true;
       expect(options.needReroll(4)).to.be.true;
     });
+  });
+
+  describe('~explodingRolls', function(){
+    it('should not have exploding rolls on by default', function(){
+      const options = new RollOptions('');
+      expect(options.explodingRolls).to.be.false;
+    });
+
+    it('should have exploding rolls if specified', function(){
+      const options1 = new RollOptions('!');
+      expect(options1.explodingRolls).to.be.true;
+
+      const options2 = new RollOptions('r3!');
+      expect(options2.explodingRolls).to.be.true;
+
+      const options3 = new RollOptions('!r3');
+      expect(options3.explodingRolls).to.be.true;
+    });
+  });
+  
+  describe('~drop', function(){
+    it('should not have drop options enabled by default', function(){
+      const options = new  RollOptions('');
+      expect(options.drop).to.be.false;
+    });
+
+    it('should have an error if no drop options are specified', function(){
+      const options = new  RollOptions('d');
+      expect(options.isValid).to.be.false;
+      expect(options.error).to.equal('Must specify options for the (d)rop modifier, e.g. d3 or dh1');
+    });
+
+    it('should have an error if drop options are invalid', function(){
+      const options = new RollOptions('dxxx');
+      expect(options.isValid).to.be.false;
+      expect(options.error).to.equal('Invalid (d)rop options: xxx');
+    });
+
+    it('should default to dropping lowest dice', function(){
+      const options = new RollOptions('d1');
+      expect(options.isValid).to.be.true;
+      expect(options.drop).to.equal(1);
+      expect(options.lowestRolls).to.be.true;
+      expect(options.highestRolls).to.be.false;
+    });
+
+    it('should allow any number of dice', function(){
+      const options1 = new RollOptions('d2');
+      expect(options1.isValid).to.be.true;
+      expect(options1.drop).to.equal(2);
+      
+      const options2 = new RollOptions('d4');
+      expect(options2.isValid).to.be.true;
+      expect(options2.drop).to.equal(4);
+    });
+
+    it('should recognize the "highest rolls" modifier', function(){
+      const options = new RollOptions('dh3');
+      expect(options.isValid).to.be.true;
+      expect(options.highestRolls).to.be.true;
+      expect(options.lowestRolls).to.be.false;
+      expect(options.drop).to.equal(3);
+    });
+  });
+    
+  describe('~keep', function(){
+    it('should not have keep options enabled by default', function(){
+      const options = new  RollOptions('');
+      expect(options.keep).to.be.false;
+    });
+
+    it('should have an error if no keep options are specified', function(){
+      const options = new  RollOptions('k');
+      expect(options.isValid).to.be.false;
+      expect(options.error).to.equal('Must specify options for the (k)eep modifier, e.g. k3 or kl1');
+    });
+
+    it('should have an error if keep options are invalid', function(){
+      const options = new RollOptions('kxxx');
+      expect(options.isValid).to.be.false;
+      expect(options.error).to.equal('Invalid (k)eep options: xxx');
+    });
+
+    it('should default to keepping lowest dice', function(){
+      const options = new RollOptions('k1');
+      expect(options.isValid).to.be.true;
+      expect(options.keep).to.equal(1);
+      expect(options.highestRolls).to.be.true;
+      expect(options.lowestRolls).to.be.false;
+    });
+
+    it('should allow any number of dice', function(){
+      const options1 = new RollOptions('k2');
+      expect(options1.isValid).to.be.true;
+      expect(options1.keep).to.equal(2);
+      
+      const options2 = new RollOptions('k4');
+      expect(options2.isValid).to.be.true;
+      expect(options2.keep).to.equal(4);
+    });
+
+    it('should recognize the "lowest rolls" modifier', function(){
+      const options = new RollOptions('kl3');
+      expect(options.isValid).to.be.true;
+      expect(options.lowestRolls).to.be.true;
+      expect(options.highestRolls).to.be.false;
+      expect(options.keep).to.equal(3);
+    });
+  });
+
+  it('should recognize a variety of options', function(){
+    const options1 = new RollOptions('!kh3r3r>17');
+    expect(options1.isValid).to.be.true;
+    expect(options1.keep).to.equal(3);
+    expect(options1.highestRolls).to.be.true;
+    expect(options1.lowestRolls).to.be.false;
+    expect(options1.explodingRolls).to.be.true;
+    expect(options1.reroll.length).to.equal(2);
+    expect(options1.needReroll(3)).to.be.true;
+    expect(options1.needReroll(18)).to.be.true;
+    expect(options1.needReroll(1)).to.be.false;
+    expect(options1.needReroll(10)).to.be.false;
+  });
+
+  it('should error if both keep and drop are specified', function(){
+    const options1 = new RollOptions('k1d1');
+    expect(options1.isValid).to.be.false;
+    expect(options1.error).to.equal('Cannot enable both "keep" and "drop" options simultaneously.');
+
+    const options2 = new RollOptions('d1k1');
+    expect(options2.isValid).to.be.false;
+    expect(options2.error).to.equal('Cannot enable both "keep" and "drop" options simultaneously.');
   });
 
 });
